requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
asn1crypto==1.5.1
bidict==0.23.1
bitarray==3.8.0
bitstring==4.3.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
esptool==5.2.0
et_xmlfile==2.0.0
eventlet==0.33.3
Flask==3.0.0
Flask-APScheduler==1.13.1
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fonttools==4.61.1
fpdf2==2.8.5
greenlet==3.3.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
influxdb==5.3.2
influxdb-client==1.49.0
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
msgpack==1.1.2
numpy==2.4.1
openai==2.32.0
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
paho-mqtt==2.1.0
pandas==2.3.3
pg8000==1.30.3
pillow==12.1.0
psutil==7.2.1
psycopg2-binary==2.9.11
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.19.2
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-engineio==4.13.1
python-pptx==1.0.2
python-socketio==5.16.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
reactivex==4.1.0
reedsolo==1.7.0
requests==2.31.0
rich==14.3.3
rich-click==1.9.7
scipy==1.17.1
scramp==1.4.8
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.0.1
wsproto==1.3.2
xlsxwriter==3.2.9
yt-dlp==2026.3.3
fastapi
uvicorn
google-generativeai
